feat: add 'Modules' solution
